LPS-57247 Move implementing classes to the proper packages
